refactor(worktree): extract IssueActionView from ActionPanel — split issue / branch right-rail flows
The right-rail panel in BranchGraph multiplexed two completely-disjoint
flows through one component via an `issueMode` prop:

- **Branch mode** (default): action surface for the selected branch —
  CI summary, PR card, delete confirmation, launch agent, focus agent,
  rebase, merge, etc. (~750 lines of JSX + handlers)
- **Issue mode** (Issues-tab focus): pre-filled "Start Work" form that
  creates a worktree and launches an agent against the selected issue,
  or surfaces an existing matching worktree.

The two never share state and don't co-render. The shared shell forced
BranchGraph's call site to thread irrelevant props through (`activeNode`
defaulted to `nodes[0]`, `prInfo` to `undefined`, `targetPrs` to
`EMPTY_PRS`, etc.) just to satisfy the type signature for the issue-mode
branch — a long-standing API smell.

This PR:

- New `IssueActionView` carries the issue-mode JSX, the start-work
  state (`startWorkName` / `startWorkBusy` / `startWorkError` /
  `startWorkInputRef`), the `buildResolvePrompt` / `handleStartWork`
  callbacks, and the `matchingWorktree` lookup. Required props shrink
  to just `selectedIssue` / `defaultBranch` / `worktrees` / `projectPath`
  / `onSelectWorktreeBranch` / `onStartWorkDone` — six instead of
  seventeen.
- `ActionPanel` drops `issueMode` / `selectedIssue` / `defaultBranch` /
  `worktrees` / `onStartWorkDone` / `onSelectWorktreeBranch` from its
  prop list, and the `if (issueMode) { return ...; }` early-out from
  its body. `WorktreeSnapshot` import + `issueToWorktreeName` import
  follow the extracted state.
- `BranchGraph`'s issues-tab call site swaps `<ActionPanel issueMode>`
  for `<IssueActionView />` — six clean props, no synthetic activeNode.

- ActionPanel.tsx 1128 → 897 (−231 lines / 20% smaller).
- IssueActionView.tsx new at 258 lines.
- 331/333 tests pass (2 pre-existing skips unrelated to this change);
  `pnpm build` clean; biome clean.
